minor enhancments
git-svn-id: http://svn.code.sf.net/p/itsucks/code/trunk@111 540001c9-3a26-0410-bb9e-87544be22eba

+++ itsucks-core/src/main/java/de/phleisch/app/itsucks/io/http/HttpMetadata.java
@@ -20,7 +20,7 @@ import de.phleisch.app.itsucks.io.Metadata;
 public class HttpMetadata extends Metadata {
 
 	private String mContentType;
-	private int mContentLength;
+	private long mContentLength;
 	private int mStatusCode;
 	private HttpMethodBase mConnection;
 	
@@ -30,10 +30,10 @@ public class HttpMetadata extends Metadata {
 	public void setContentType(String mimetype) {
 		mContentType = mimetype;
 	}
-	public int getContentLength() {
+	public long getContentLength() {
 		return mContentLength;
 	}
-	public void setContentLength(int contentLength) {
+	public void setContentLength(long contentLength) {
 		mContentLength = contentLength;
 	}
 	public int getStatusCode() {
